Corrected Organization redirect to differentiate between updated and new.
Previously, the Organization page would redirect you to the homepage
regardless of whether you were creating a new organization or editing
an existing organization. Now it redirects to the Zone page if you
have just created a new organization, and back to the Homepage if
you were merely editing an existing organization.