Styled Facility Admin page and added link to it from Facility Selection.
The Facility Admin page needed to be styled such that the facility
names and buttons cleanly aligned regardless of word length. In addition,
the site previously did not have any way of navigating to the Facility
Admin page. Therefore a link was added to it from the Facility Selection
page.