Fixed bug regarding topic links.
The home page previously displayed a set list of top-level topics,
yet one of these topics has since been moved to a different level
in the topic-tree. The homepage template was changed such that topic
titles are no longer hard coded, but instead populate based on the top-
level topics in topics.json.