Added option to launch mplayer in the background to play videos.
The Raspberry Pi browsers cannot yet play video. But mplayer works on the
Pi, so we can trigger it to be launched when the user wants to view a
video. Point tracking is done in the backend Python thread that launched
mplayer, and the browser refreshes the point total when mplayer is closed.

Add `USE_MPLAYER = True` to local_settings.py to enabled this. Note that
it will only work when the browser is on the same computer as the server.