Handle failures in setting copied script file permissions gracefully.
In some cases, KA Lite is installed with read/write permissions, but
not owned by the user who will be running it (e.g. in a shared environ).
In this case, the setup script would fail when trying to set the stats
of a copied script, and end up in an unrecoverable state ("buddy, you're
screwed").

Since it's likely to work fine without setting the file stats, this PR
switches to just warning, rather than dying.