translate exercise data with whatever language was requested.
Before, we translated based on exercise_lang, whose value is whatever
get_best_available_language returns for a given set of
languages. However, those set of language choices we pass in are those
we have translated exercise templates for. When translating for perseus,
this is not the case.

So what we do here instead is just localize for whatever language we
originally requested for get_exercise_cache in the first place. If it's
not there, gettext automatically falls back to English.