don't load the settings during function eval.
This causes a circular import where a settings module imports this
module at the toplevel. Doing that import forces django to load the
LazySettings attributes, which causes an import of all settings modules,
which causes this module to get loaded, which forces django to load the
LazySettings attributes...

Loading settings at actual function call avoids this circular import.