Add "before-showing" logic to ButtonView
Attempts to stay close to the spec, but with one caveat:
Bootstrap .dropdown elements can be clicked but would then be collapsed, presumably by
the internal logic of bootstrap. To circumvent this, in any given step, if an element is clicked
_and_ if the target element of the introJs step (these elements could be distinct) has a parent
with the bootstrap .dropdown class, then the bootstrap dropdown-api will be disabled after the
click, an re-enabled once the user exits or completes the intro.

The caveat is that if another _subsequent_ step depends on clicking a dropdown menu, then since
the api will have been disabled, it will not function correctly.