Add more explicit waits to steps.py file
In the case of the "next steps" view, since it loads subviews, it avoids a potential race condition to
explicitly wait for an element within the subviews. Otherwise the parent view can load, be found by the
webdriver, and the child element searched for before the child element has loaded.

In the case of the other two views, this consideration doesn't apply since they don't contain subviews.
But in general, if you want to find a particular element, why not give it a name -- either through the id
or class attributed? This makes the test safe against changes to the layout.