Increase max wait time for behave_helpers.py
Seems like whenever a change is made on the front end, another unrelated test starts timing out.
Why is the app so dang slow under automation? Is it slow in fact, or is it a problem with the
test environment?

Being unable to answer these questions, I've increased the max timeout across the board, in the
hope of preventing unrelated timeout errors down the road. This could potentially cause the tests
to take signficiantly longer to run.