Create basic functionality for star rating view
Each star rating has a number of elements with class .star-rating-option, each of which requires
a "data-val" property that defines the value of the rating when you click that element.
"data-val" should be an integer.

When a .star-rating-option is clicked, then the rating attribute of the StarView's model is
updated, and all .star-rating-options with lesser or equal value are given the css class "activated".

The model of a StarView may be set through the constrcutor options. The name of the model attribute
corresponding to the rating may also be set with the "rating_attr" option, and defaults to "rating".

The intention is that StarViews can share their model with a parent view, and each star_view will
control a different attribute, e.g. "xyz-rating" and "abc-rating", of the parent model.