Create the basic user experience for rating content
If a rating exists (if the model is fetched successfully) then all the data will be displayed on
a completed form -- meaning all star rating views will be visible, as well as the text, with an edit btn.

If the rating does not exist, the user will be prompted to fill in each star rating and the text,
one by one, then finally all the user's data will be displayed just as though it was fetched successfully.

Because the model's url is stubbed out with a fake url, I address an error in the failure callback.
Since it's possible for the response text to contain html, we just don't attempt to render the response
text in the alert bubble.