Use less eager redirection
Redirecting all requests was too eager... some requests should be unimpeded.
One idea is to intercept the response -- by looking at content-type, I hope
to intercept all html responses, but let javascript or css through.

It might also be possible to filter requests on it's content-type, it's path
or some other characteristic, through a combination of whitelisting and
blacklisting.