Change version comparison logic
The central server sends a short version (e.g. 0.14), which the distributed server
then compares to its long version (e.g. 0.14.1). The problem is that 0.14 compares
as less than 0.14.1, so language packs are shown as unavailable. Instead,
when comparing two mismatched-length versions, only consider the shortest version
available -- thus 0.14 and 0.14.1 would compare equal. Then the languagepack versions
will be compared, which is the actual intent.