Reel in odd test
Let it be sufficient that the highlighted class appears in the DOM.

Plus, if the intro is started before all the handlebars templates finish rendering,
we're in trouble -- elements may not exist. Since there's no reliable way to
indicate that the page is done loading all handlebars templates, let's not even
go down that rabbit hole.