Refactor behave_helper functions
DRY them out. Remove potential error -- if an element is not found initially, then
the assert_no_element_by_* functions will pass. But we should try to find the element
with a wait, in order to catch the case where it doesn't appear initally but *does*
appear some time after the page loads. *Then* we wait for it to get stale.