In learn_content, wait for page to load; refactor helper
_assert_no_element_by refactored -- previously we were seemingly
trying to handle the case where an element didn't exist on the page,
then *did* exist, then became stale. That's a difficult general condition
to encapsulate, so instead let's just check that the element doesn't exist.

Then it becomes the caller's responsibility to ensure that the element has
sufficient time to appear, if it will appear at all. We ensure that in
the learn_content test by waiting for ajax stuff to wrap up.