In initialize_kalite, close db connection
We determine whether to run setup by checking the database, thus
opening a database connection. With sqlite, that means file access.
In that case we'll call setup, and setup will try move the database
file, but it can't (at least on Windows) because the file is open.
So close the database connection (thus releasing the file) before
calling setup. Ta-da!