Address rating responsive issues; fixes #4544
Played around with different layouts within bootstrap,
not totally satisfying for resolutions > ~400px and < 768 px,
but good enough without changing the hb template or view logic.

At smallest res, all rating buttons and labels stack,
and no hover-label is shown.

As the resolution increases (to bootstrap's col-sm- size),
the labels and buttons flatten out, and a hover label is
shown. The size of the buttons also progressively increases.