Refactor export view, fix bug where Facility selector is disabled
I dug through the commit messages to determine that the intent was
to disable the Facility selector when the requesting user is a
FacilityUser (not an admin). Instead of a conditional in the
View's render function, I tried to make the intent more explicit
by exposing a new context variable and putting the disable logic
in the View's is_disabled function.

Also fixes a bug where the Facility selector is erroneously
disabled when some resources are selected.