Allow sidebar to finish rendering when no content is found
No content could be found if the content is unavailable, in which case our
api endpoint will return nothing. It's also possible if a path is entered in the
address bar which doesn't correspond to an existing content item.

The ultimate aim is to allow the page to finish loading gracefully, even when
such errors are encountered.

Could also help in creating "test" content items for the test framework.