Ensure kalite is initialized in base mgmt command
The class `UpdatesCommand` is the base class for
the `retrievecontentpack` command. Since this could
potentially be run *before* setup, it produces an error
when the database is accessed by the `UpdateProgressLog`
Model methods -- the database wouldn't exist!

`initialize_kalite` ensures that no such error is raised
by creating a database.