Force clear out collectstatic on setup
In some upgrade scenarios (like on the RPi) the system's
built js files might have a timestamp *earlier* than the
static files appearing in the user's static files location.
The effect is that `collectstatic` called during an upgrade
will not replace the old static files with the new ones, since
the timestamp is reverse.

So force them to be cleared. But preserve the "extra" files we
add in, e.g. for subtitles (vtt, srt) and exercise templates (html).
Those must not be removed, since they're downloadable content!